ht1/Consumer/requirements.txt
pika
psycopg2-binary
datetime
